+++ examples/basic.rs
@@ -32,10 +32,11 @@ async fn get_user(Path(id): Path<u64>) -> Json<UserResponse> {
     })
 }
 
-fn main() {
+#[tokio::main]
+async fn main() {
     fastrs::App::new()
         .route(create_user)
         .route(get_user)
         .serve_docs_at("/docs")
-        .run("0.0.0.0:8000");
+        .run("0.0.0.0:8000").await;
 }